Working solution on all nav areas
