+++ src/components/sideNav/SideNavEntry.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-base-to-string */
 import * as React from 'react';
 
 import classNames from 'classnames';
@@ -6,7 +5,7 @@ import { Location, LocationDescriptor } from 'history';
 import { match as Match } from 'react-router';
 import { NavHashLink, NavHashLinkProps } from 'react-router-hash-link';
 import './SideNav.scss';
-import { history } from '../../store';
+import { isNavHashLinkExact } from '../../utils/isNavHashLinkExact';
 
 export interface SideNavEntryProps extends NavHashLinkProps {
   name: string | JSX.Element;
@@ -70,24 +69,6 @@ const SideNavEntry = (props: SideNavEntryProps): JSX.Element => {
     }
   };
 
-  const navHashLinkIsExact = (
-    to:
-      | LocationDescriptor<unknown>
-      | ((location: Location<unknown>) => LocationDescriptor<unknown>),
-  ): boolean => {
-    const url = to.toString();
-    if (url.startsWith('#')) {
-      // On a hash link match to just the hash because pathname isn't present
-      return url === history.location.hash;
-    } else {
-      /*
-       * On a regular link match to the full pathname and hash to
-       * precent parent links from getting aria-current="page"
-       */
-      return url === history.location.pathname + history.location.search + history.location.hash;
-    }
-  };
-
   // Omit unneeded parent props from NavLink
   /* eslint-disable @typescript-eslint/no-unused-vars -- omit sharedAnchors from navLinkProps */
   const { name, className, subNavLevel, sharedAnchors, forceAriaCurrent, ...navLinkProps } = props;
@@ -117,7 +98,7 @@ const SideNavEntry = (props: SideNavEntryProps): JSX.Element => {
           'vads-u-border-left--5px': subNavLevel === 0,
         })}
         isActive={navHashLinkIsActive}
-        aria-current={props.forceAriaCurrent || navHashLinkIsExact(props.to) ? 'page' : 'false'}
+        aria-current={props.forceAriaCurrent || isNavHashLinkExact(props.to) ? 'page' : 'false'}
         {...navLinkProps}
       >
         {name}
